requirements.txt
aiobotocore==1.4.2
aiohttp==3.7.4.post0
aioitertools==0.8.0
async-timeout==3.0.1
attrs==21.2.0
boto3==1.19.4
botocore==1.20.106
chardet==4.0.0
cycler==0.11.0
idna==3.3
jmespath==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
multidict==5.2.0
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.3
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.13.2
yarl==1.7.0
